+++ src/structs/Livestream.jl
@@ -9,6 +9,7 @@ Holds the conguration for livestream, defaults to `nothing`
 - `address::String` The IP address for the `:local` stream(ignored in case of `:twitch`)
 - `port::Int` The port for the `:local` stream(ignored in case of `:twitch`)
 - `twitch_key::String` Twitch stream key for your account
+- `frames::Union{UnitRange{Int}, Symbol}` The specific frames to be livestreamed. Defaults to `:all` 
 """
 struct StreamConfig
     livestreamto::Symbol
@@ -16,4 +17,5 @@ struct StreamConfig
     address::String
     port::Int
     twitch_key::String
+    frames::Union{UnitRange{Int},Symbol}
 end